feat(tools): 实现工具自动注册和管理功能
- 在AbstractTool中添加ToolRegistry类，支持异步加载所有工具模块
- 实现工具的自动发现、注册、禁用和重新加载功能
- 添加线程安全的锁机制确保并发安全
- 统一通过registry进行工具管理和访问
- 移除原有的装饰器注册方式，改为自动扫描注册
- 优化工具目录结构和导入路径
